workshop-rag/backend-functions/requirements.txt
azure-functions
azure-functions-openai
